feat: add baselayout w/ markdown & default layout variants
this is needed to make the layout & scrolling work as intended for
markdown rendered pages as well as custom ones which don't necessarily
fill a full page.

Maybe this can also be achieved with smarter css, e.g. just setting the
main section to the screen height and applying a top heading or so, but
I didn't figure it out right away.